src/backend/requirements.txt
# FastAPI and Server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.2

# AI/ML Models
torch==2.5.1
transformers==4.46.3
accelerate==1.1.1

# Audio Processing
librosa==0.10.2
soundfile==0.12.1
numpy==1.26.4

# Additional utilities
python-dotenv==1.0.1

# Google Cloud Vertex AI
google-cloud-aiplatform>=1.70.0
google-genai>=0.2.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
